test(xcompare): move CLI exec tests to Integration suite
